Create Server_With_Three_Separate_Endpoints.go
The Server is a three Separate Endpoints server. The homepage Endpoint with a title, and an animated gif file.
The Post Endpoint has an appended slice, and a structure to which, the fields of an instance of the struct that has been instantiated.
The Get endpoint simply returns the slice and structure values.